requirements.txt
# Copyright 2017 Intel Research and Development Ireland Limited
# *************************************************************

# This file is part of OSM Monitoring module
# All Rights Reserved to Intel Corporation

# Licensed under the Apache License, Version 2.0 (the "License"); you may
# not use this file except in compliance with the License. You may obtain
# a copy of the License at

#         http://www.apache.org/licenses/LICENSE-2.0

# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations
# under the License.

# For those usages not covered by the Apache License, Version 2.0 please
# contact: prithiv.mohan@intel.com or adrian.hoban@intel.com
kafka-python==1.4.*
requests==2.18.*
cherrypy==14.0.*
jsmin==2.2.*
jsonschema==2.6.*
python-keystoneclient==3.15.*
boto==2.48
python-cloudwatchlogs-logging==0.0.3
py-cloudwatch==0.0.1
pyvcloud==19.1.1
pyopenssl==17.5.*
six==1.11.*
bottle==0.12.*
peewee==3.1.*
pyyaml==3.*
prometheus_client==0.4.*
gnocchiclient==7.0.*
git+https://osm.etsi.org/gerrit/osm/common.git#egg=osm-common
git+https://osm.etsi.org/gerrit/osm/N2VC.git#egg=n2vc